backend/requirements.txt
fastapi>=0.128.0
langchain>=1.2.2
langchain-community>=0.4.1
langchain-openai>=1.1.7
langchain-pinecone>=0.2.13
langchain-text-splitters>=1.1.0
langgraph-checkpoint-postgres>=3.0.3
langgraph[postgres]>=1.0.5
pinecone-client>=6.0.0
psycopg[binary,pool]>=3.3.2
pydantic-settings>=2.0.0
pypdf>=6.5.0
python-dotenv>=1.2.1
python-multipart>=0.0.21
uvicorn>=0.40.0
pyjwt>=2.8.0
google-auth>=2.27.0
httpx>=0.26.0
